Added info on how to use the DDK Examples with the earlier DDK v1.1
+ Updated the README with the relevant information. With this the latest
revision of Examples tree, could be used with all DDK variants, with
appropriate flags, as explained in the README.
+ To achieve this multi-version support, the key change required was in the
commandline of the bootloadHID, i.e. to support downloading the hex code to
different DDK (bootloader) versions based on the options passed to it.
+ Once done with that change, the appropriate options have been actually
passed in the invocation of the bootloadHID commandline through the relevant
make files (Examples/Makefile, rules.mk).